\name{compSPOT}
\alias{compSPOT}

\title{
Mutation hotspot finder and comparison of hotspot burden.
}
\description{
It is well known that numerous clones of cells sharing common
mutations exist within cancer, precancer, and even clinically
normal appearing tissues. The frequency and location of
these mutations may aid in the prediction of cancer risk of
certain individuals. It has also been well established that
certain genomic regions have increased sensitivity to acquiring
mutations. Mutation-sensitive genomic regions may therefore be
used as markers for prediction of cancer risk. This package
contains multiple functions for the establishment of significantly
mutated hotspots, comparison of hotspot mutation burden between
sub-groups, and exploratory data analysis of the correlation between
hotspot mutation burden, and personal risk factors for cancer such
as age, gender, and history of carcinogen exposure. This package
aims to allow users to identify robust genomic markers
which may serve as markers of cancer risk.
}
\value{
A package containing functions which find statistically significant
mutation hotspots and compare mutation hotspot burden between groups
and correlation between clinical features.
}

