% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.spot.R
\name{feature.spot}
\alias{feature.spot}
\title{hotspot comparison by additional features}
\usage{
feature.spot(data, regions, feature)
}
\arguments{
\item{data}{A dataframe containing the clinical features and the mutation count.}

\item{features}{A dataframe containing all the features.}
}
\value{
A grid of all the violin plots for the categorical data and scatter plot for the sequential data.
}
\description{
This function classifying the features into categorical and sequential features
and return a grid of violin plots for categorical features and scatter plot for sequential features.
}
\details{
This function is used to classify the features into sequential features if values are integers or classifies them into categorical #' features.
Sequential features are compared to the mutation count using Pearson correlation.
Similarly, in categorical features wilcox test is used to compare between the groups in the features based on their mutational
count.
Scatter plot is used to represent the sequential features along with the R and p-value from the pearson correlation.
Similarly violin plots are used to plot the groups in the categorical data and wilcox or kruska-wallis values are shown on the
graph.
}
\examples{

data("example_mutations")
data("example_regions")
features <- c("AGE", "SEX", "ADJUVANT_TX", "SMOKING_HISTORY", "TUMOR_VOLUME", "KI_67")
feature.spot(data = example_mutations, regions = example_regions, feature = features)

}
