% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.spot.R
\name{group.spot}
\alias{group.spot}
\title{hotspot comparison by group}
\usage{
group.spot(data, regions, pvalue, threshold, name1, name2, include_genes)
}
\arguments{
\item{data}{a dataframe containing the chromosome, base pair position, and optionally gene name of each mutation}

\item{regions}{a dataframe containing the chromosome, start and end base pair position of each region of interest}

\item{threshold}{the cutoff empirical distribution for Kolmogorov-Smirnov test}

\item{name1}{a string containing the name of one group for the comparison}

\item{name2}{a string containing the name of the second group for the comparison}

\item{pval}{a threshold p-value for Kolmogorov-Smirnov test}
}
\value{
A dataframe containing the genomic regions with significant mutation frequency and summary plots for
mutation comparison between groups
}
\description{
This function compares the mutation frequency of a panel of genomic regions between two sub-groups.
}
\details{
This function creates a list of mutation frequency per unique sample for each genomic regions separated
based on specified sub-groups. The regions with significant differences in mutation distribution are
calculated using a Kolmogorov-Smirnov test. The difference in mutation frequency is output in a violin plot.
}
\examples{

data("example_mutations")
data("example_regions")
group.spot(data = example_mutations, regions = example_regions, pval = 0.05, threshold = 0.4,
name1 = "High-Risk", name2 = "Low-Risk", include_genes = TRUE)

}
