% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.spot.R
\name{plot.sigspot}
\alias{plot.sigspot}
\title{sig.spot visualization}
\usage{
\method{plot}{sigspot}(df1, df2)
}
\arguments{
\item{df1}{a dataframe containing the labeled region number, type of region, percentage of samples with
hotspot mutation}

\item{df2}{a dataframe containing the data required for calculation of ECDF}
}
\value{
A figure organizing the region number in ranked order versus the percentage of the sample with hotspot
mutation which also denotes the label of the data point is displayed when hovering over the item. A second
figure providing a visualization of the ECDF for hotspot vs non-hotspots.
}
\description{
Based on a panel of genomic regions, this function visualizes the regions which are found to have
significance based on percentage of samples with hotspot mutations.
}
\details{
This function begins by displaying a plot with a ranked number of hotspot based on number of mutations
on the x-axis versus percentage of all samples with mutations in hotspot on the y-axis. Regions of high
significance are ranked on the left side of the plot. Based on the type, the points plotted as red
dots are the hotspots which were found to have a significant number of mutations whereas the pink were
found to not be significant. An interactive label of each point is also available. A second plot is also
created using the same data used in the Kolmogorov-Smirnov calculation to show a visualization of the ECDF
of hotspots vs non-hotspots.
}
\examples{

data("regions")
data("ecdf_df")
plot <- plot.sigspot(df1 = region, df2 = ecdf_df)

}
\keyword{internal}
