% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.spot.R
\name{plot.spot}
\alias{plot.spot}
\title{hotspot group comparison with plot}
\usage{
\method{plot}{spot}(plot_data, group1, group2)
}
\arguments{
\item{plot_data}{a dataframe containing the hotspot, mutation count, and associated groups}

\item{group1}{a string containing the name of one group for the comparison}

\item{group2}{a string containing the name of the second group for the comparison}
}
\value{
An interactive violin plot displaying the comparison of mutation counts for each group.
}
\description{
Based on a dataframe of hotspots, this function groups the hotspots into two different groups and overlays interactive
violin plots based on mutation count for comparison.
}
\details{
This function compares the mutation frequency per sample for each hotspot in a violin plot.
}
\examples{

data("plot_data")
fig <- plot.spot(plot_data = plot_data, group1 = "High-Risk", group2 = "Low-Risk")

}
\keyword{internal}
