% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.spot.R
\name{sig.spots}
\alias{sig.spots}
\title{significant hotspot calculator}
\usage{
sig.spots(data, regions, pvalue, threshold, include_genes, rank)
}
\arguments{
\item{data}{a dataframe containing the chromosome, base pair position, and optionally gene name of each mutation}

\item{regions}{a dataframe containing the chromosome, start and end base pair position of each region of interest}

\item{pvalue}{the p-value cutoff for included hotspots}

\item{threshold}{the cutoff empirical distribution for Kolmogorov-Smirnov test}

\item{include_genes}{true or false whether gene names are included in regions dataframe}

\item{rank}{true or false whether regions dataframe is already ranked and includes mutation count of total dataset}
}
\value{
A dataframe containing the genomic regions with significant mutation frequency and plots summarizing
significant hotspots
}
\description{
Based on a panel of genomic regions, this function calculates the regions which are found to have
significantly higher mutation frequency compared to less mutated regions.
}
\details{
This function begins by measuring the mutation frequency for each unique sample for each provided genomic region.
Beginning with the top ranked hotspot, a Kolmogorov-Smirnov test is preformed on the mutation frequency of the
top genomic region compared to the normalized mutation frequency of all the lower-ranked regions. This
continues, then running the Kolmogorov-Smirnov test for the normalized mutation frequency of the top 2 genomic
regions compared to the normalized mutation frequency of all lower-ranked regions.This process repeats itself,
continuously adding an additional genomic regions each time until either the set p-value or empirical distribution
threshold is not met. Once this cutoff has been reached, an established list of mutation hotspots is provided.
}
\examples{

data("example_mutations")
data("example_regions")
significant_spots <- sig.spots(data = example_mutations, regions = example_regions,
pvalue = 0.05, threshold = 0.2, include_genes = TRUE, rank = TRUE)


}
