% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.spot.R
\name{spot.rank}
\alias{spot.rank}
\title{genomic region ranker}
\usage{
spot.rank(data, regions, include_genes)
}
\arguments{
\item{data}{A dataframe containing the location of each mutation.}

\item{regions}{a dataframe containing the location of gene region of interest}

\item{include_genes}{true or false whether gene names are included in regions dataframe}
}
\value{
A dataframe containing the ranked genomic regions from highest to lowest mutation frequency
}
\description{
ranks all genomic regions from highest to lowest mutation frequency in given dataset
}
\examples{

data("mutation_data")
spot.rank(data = mutation_data, regions = ea_regions, include_genes = TRUE)

}
\keyword{internal}
